Always preserve prov's config.yaml
Also, if for whatever reason it's not there, make sure to not overwrite
the one from write-config step.

A grace period needs to be explicitly defined for this to work.

+++ ci-operator/step-registry/aro-hcp/provision/environment/aro-hcp-provision-environment-commands.sh
@@ -60,15 +60,23 @@ yq eval -n "
 echo "Created override config at: ${OVERRIDE_CONFIG_FILE}"
 cat ${OVERRIDE_CONFIG_FILE}
 
+CONFIG_PROV="${SHARED_DIR}/config-prov.yaml"
+
+finalize() {
+    if [[ -s "${CONFIG_PROV}" ]]; then
+        mv "${CONFIG_PROV}" "${SHARED_DIR}/config.yaml"
+        cp "${SHARED_DIR}/config.yaml" "${ARTIFACT_DIR}/config.yaml"
+    fi
+}
+trap finalize EXIT
+
 unset GOFLAGS
 make -o tooling/templatize/templatize entrypoint/Region \
   DEPLOY_ENV=prow \
   EXTRA_ARGS="--region ${LOCATION} --abort-if-regional-exist" \
   TIMING_OUTPUT=${SHARED_DIR}/steps.yaml.gz \
   ENTRYPOINT_JUNIT_OUTPUT=${ARTIFACT_DIR}/junit_entrypoint.xml \
-  CONFIG_OUTPUT=${SHARED_DIR}/config.yaml
-
-cp "${SHARED_DIR}/config.yaml" "${ARTIFACT_DIR}/config.yaml"
+  CONFIG_OUTPUT=${CONFIG_PROV}
 
 # Mark successful completion
 touch "${SHARED_DIR}/provision-complete"
